@@ -7,7 +7,7 @@ readme = "README.md"
 packages = [{include = "nbutils", from = "src"}]
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.12"
 click = "^8.1.7"
 nbformat = "^5.9.2"
 
