@@ -1,15 +1,19 @@
+"""Core NBUtils class for notebook operations."""
+
 import json
-import os
-from typing import Optional
+from pathlib import Path
 
 from .operations.convert import notebook_to_markdown, notebook_to_py, py_to_notebook
-from .operations.headings import adjust_headings
+from .operations.headings import adjust_file_headings
+
 
 class NBUtils:
-    def __init__(self, input_path: Optional[str] = None):
+    """Main utility class for notebook operations."""
+    
+    def __init__(self, input_path: str | None = None):
         self.input_path = input_path
         
-    def convert_to_markdown(self, output_path: Optional[str] = None) -> None:
+    def convert_to_markdown(self, output_path: str | None = None) -> str:
         """Convert notebook to markdown"""
         with open(self.input_path, 'r', encoding='utf-8') as f:
             notebook_content = json.load(f)
@@ -21,7 +25,7 @@ class NBUtils:
                 f.write(markdown_content)
         return markdown_content
 
-    def convert_to_py(self, output_path: Optional[str] = None) -> None:
+    def convert_to_py(self, output_path: str | None = None) -> str:
         """Convert notebook to Python file"""
         with open(self.input_path, 'r', encoding='utf-8') as f:
             notebook_content = json.load(f)
@@ -33,11 +37,18 @@ class NBUtils:
                 f.write(py_content)
         return py_content
 
-    def adjust_headings(self, adjustment: str) -> None:
-        """Adjust heading levels in the notebook"""
-        adjust_headings(self.input_path, adjustment)
+    def adjust_headings(self, increase: bool, force: bool = False) -> None:
+        """Adjust heading levels in the notebook.
         
-    def convert_from_py(self, output_path: Optional[str] = None) -> None:
+        Args:
+            increase: True to increase heading levels, False to decrease
+            force: If True, proceed even with first-level headings when decreasing
+        """
+        result = adjust_file_headings(self.input_path, increase, force)
+        if not result.success:
+            raise ValueError(result.error or "Failed to adjust headings")
+        
+    def convert_from_py(self, output_path: str | None = None) -> dict[str, any]:
         """Convert Python file to Jupyter notebook"""
         with open(self.input_path, 'r', encoding='utf-8') as f:
             py_content = f.read()
